@@ -1,16 +1,17 @@
 name: parallax_bg
 description: Create beautiful & interactive parallax backgrounds
-version: 1.0.4
+version: 2.0.0
 homepage: https://github.com/kumar-aakash86/parallax_bg
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_sensors: ^0.1.6
+  sensors_plus: ^1.2.1
+
 
 dev_dependencies:
   flutter_test:
